sql: Move SplitTable into TestCluster
refs https://github.com/cockroachdb/cockroach/issues/85800

Move this function to make it accessible by the TTL (and other) tests.

Release note: None

Release justification:
