+++ pkg/sql/distsql_physical_planner_test.go
@@ -59,80 +59,6 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-// SplitTable splits a range in the table, creates a replica for the right
-// side of the split on TargetNodeIdx, and moves the lease for the right
-// side of the split to TargetNodeIdx for each SplitPoint. This forces the
-// querying against the table to be distributed.
-//
-// TODO(radu): SplitTable or its equivalent should be added to TestCluster.
-//
-// TODO(radu): we should verify that the queries in tests using SplitTable
-// are indeed distributed as intended.
-func SplitTable(
-	t *testing.T, tc serverutils.TestClusterInterface, desc catalog.TableDescriptor, sps []SplitPoint,
-) {
-	if tc.ReplicationMode() != base.ReplicationManual {
-		t.Fatal("SplitTable called on a test cluster that was not in manual replication mode")
-	}
-
-	rkts := make(map[roachpb.RangeID]rangeAndKT)
-	for _, sp := range sps {
-		pik, err := randgen.TestingMakePrimaryIndexKey(desc, sp.Vals...)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		_, rightRange, err := tc.Server(0).SplitRange(pik)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		rightRangeStartKey := rightRange.StartKey.AsRawKey()
-		target := tc.Target(sp.TargetNodeIdx)
-
-		rkts[rightRange.RangeID] = rangeAndKT{
-			rightRange,
-			serverutils.KeyAndTargets{StartKey: rightRangeStartKey, Targets: []roachpb.ReplicationTarget{target}}}
-	}
-
-	var kts []serverutils.KeyAndTargets
-	for _, rkt := range rkts {
-		kts = append(kts, rkt.KT)
-	}
-	descs, errs := tc.AddVotersMulti(kts...)
-	for _, err := range errs {
-		if err != nil && !testutils.IsError(err, "is already present") {
-			t.Fatal(err)
-		}
-	}
-
-	for _, desc := range descs {
-		rkt, ok := rkts[desc.RangeID]
-		if !ok {
-			continue
-		}
-
-		for _, target := range rkt.KT.Targets {
-			if err := tc.TransferRangeLease(desc, target); err != nil {
-				t.Fatal(err)
-			}
-		}
-	}
-}
-
-// SplitPoint describes a split point that is passed to SplitTable.
-type SplitPoint struct {
-	// TargetNodeIdx is the node that will have the lease for the new range.
-	TargetNodeIdx int
-	// Vals is list of values forming a primary key for the table.
-	Vals []interface{}
-}
-
-type rangeAndKT struct {
-	Range roachpb.RangeDescriptor
-	KT    serverutils.KeyAndTargets
-}
-
 // TestPlanningDuringSplits verifies that table reader planning (resolving
 // spans) tolerates concurrent splits and merges.
 func TestPlanningDuringSplitsAndMerges(t *testing.T) {
