+++ pkg/testutils/testcluster/BUILD.bazel
@@ -17,6 +17,8 @@ go_library(
         "//pkg/rpc/nodedialer",
         "//pkg/server",
         "//pkg/server/serverpb",
+        "//pkg/sql/catalog",
+        "//pkg/sql/randgen",
         "//pkg/storage",
         "//pkg/testutils",
         "//pkg/testutils/serverutils",
